Move protractor stuff to a folder
