+++ packages/perseus/src/widgets/interactive-graphs/mafs-graph.tsx
@@ -31,7 +31,7 @@ import {
 import {interactiveGraphReducer} from "./reducer/interactive-graph-reducer";
 import {getGradableGraph, getRadius} from "./reducer/interactive-graph-state";
 import {GraphConfigContext} from "./reducer/use-graph-config";
-import {StatefulProtractor} from "./stateful-protractor";
+import {StatefulProtractor} from "./protractor/stateful-protractor";
 
 import type {InteractiveGraphState, InteractiveGraphProps} from "./types";
 import type {PerseusGraphType} from "../../perseus-types";
